+++ frontend/src/service/auth-service.ts
@@ -1,5 +1,19 @@
-import {catchError, map, Observable, ReplaySubject, tap, throwError} from "rxjs";
+import {
+    catchError,
+    interval,
+    map,
+    noop,
+    Observable,
+    of,
+    ReplaySubject,
+    takeWhile,
+    tap,
+    throwError
+} from "rxjs";
 import AuthApiService, {DefaultAuthApiService} from "../api/auth-api-service";
+import {first} from "rxjs/operators";
+
+const REFRESH_INTERVAL = 1000 * 60 * 15; // 15 mins
 
 export default class AuthService {
 
@@ -8,10 +22,26 @@ export default class AuthService {
 
     constructor(private authApiService: AuthApiService = new DefaultAuthApiService()) {
         this.subject.next(false);
+        interval(REFRESH_INTERVAL).pipe(takeWhile(ev => this.token !== 'authorized')).subscribe(() => this.refresh().pipe(first()).subscribe(noop));
     }
 
     authChannel(): Observable<boolean> {
         return this.subject;
+    }
+
+    refresh(): Observable<boolean> {
+        if (this.token) {
+            return this.authApiService.refresh().pipe(map(auth => {
+                this.token = auth.token;
+                return auth.token != null
+            }), tap(data => {
+                this.subject.next(data);
+            }), catchError((error: any) => {
+                this.subject.next(false);
+                return throwError(() => error)
+            }));
+        }
+        return of(false);
     }
 
     authenticate(username: string, password: string): Observable<boolean> {
@@ -20,17 +50,14 @@ export default class AuthService {
             return auth.token != null
         }), tap(data => {
             this.subject.next(data);
-        }), catchError((error:any) => {
-            this.subject.next(false);
-            return throwError(() => error)
-        }));
+        }), catchError(() => of(false)));
     }
 
     isAuthenticated(): boolean {
         return this.token != null;
     }
 
-    getToken(): string  {
+    getToken(): string {
         return this.token;
     }
 }